WIP: Move components to core or features
- Adds clarity between different components.
